requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
streamlit
plotly-express